// src/types/config.ts
export interface FullscreenWallpaperConfig {
	enable?: boolean;
	src:
		| string
		| string[]
		| {
				desktop?: string | string[];
				mobile?: string | string[];
		  };
	position?: "top" | "center" | "bottom";
	carousel?: {
		enable: boolean;
		interval: number;
	};
	zIndex?: number;
	opacity?: number;
	blur?: number;
	switchable?: boolean;
	overlay?: {
		opacity?: number;
		blur?: number;
		cardOpacity?: number;
		switchable?:
			| boolean
			| {
					opacity?: boolean;
					blur?: boolean;
					cardOpacity?: boolean;
			  };
	};
	fullscreen?: {
		switchable?:
			| boolean
			| {
					opacity?: boolean;
					blur?: boolean;
			  };
	};
}